requirements.txt
Flask==3.0.0
requests
python-dotenv
Werkzeug
flask-cors
gunicorn